+++ lib/model/review/review_model.dart
@@ -7,7 +7,7 @@ part 'review_model.g.dart';
 @JsonSerializable()
 class Review {
   int id;
-  int? rating;
+  double? rating;
   String? review;
   UserModel? user;
   CourseModel? course;
